General-> final version before testing

+++ database/migrations/2023_09_21_113431_create_carts_table.php
@@ -16,12 +16,12 @@ return new class extends Migration
       $table->string('name');
       $table->string('session_id');
       $table->integer('quantity_amount')->default(0);
-      $table->integer('sum_amount')->default(0);
+      $table->double('sum_amount')->default(0);
       $table->unsignedBigInteger('currency_id')->index()->nullable();
       $table->foreign('currency_id')->references('id')->on('currencies');
       $table->unsignedBigInteger('status_id')->index()->nullable();
       $table->foreign('status_id')->references('id')->on('statuses');
-      $table->integer('final_amount')->default(0);
+      $table->double('final_amount')->default(0);
       $table->string('order_id')->nullable();
       $table->timestamps();
     });
